+++ Changelog.md
@@ -5,6 +5,7 @@
 - button to cleanup localStorage invalid entries
 - Link object, major rewrite ensues
 - LibraryItem params conf, reduces or eliminates builder function necessity
+- select param can have a datasource to automaticaly add / remove items
 
 ## pre 2014-08-06